Extract separate class for external DTD options tests
